+++ src/components/ui/index.tsx
@@ -1,0 +1,4 @@
+export {default as button} from "./button";
+export {default as Card} from "./Card";
+export {default as ContactForm} from "./ContactForm";
+export {default as FeatureGrid} from "./FeatureGrid";
