+++ src/components/variants/TitleVariants.ts
@@ -1,0 +1,13 @@
+const titleVariants = {
+    hidden: { opacity: 0, y: -20 },
+    visible: { 
+      opacity: 1, 
+      y: 0,
+      transition: {
+        duration: 0.5,
+        staggerChildren: 0.1
+      }
+    }
+};
+
+export default titleVariants;
