+++ src/components/variants/WordVariants.ts
@@ -1,0 +1,14 @@
+const wordVariants = {
+    hidden: { opacity: 0, y: 20 },
+    visible: { 
+      opacity: 1, 
+      y: 0,
+      transition: {
+        type: 'spring',
+        damping: 12,
+        stiffness: 200
+      }
+    }
+};
+  
+export default wordVariants;
