+++ src/components/variants/index.ts
@@ -1,0 +1,2 @@
+export {default as titleVariants} from "./TitleVariants";
+export {default as wordVariants} from "./WordVariants";
